load/requirements.txt
icrawler
selenium
webdriver-manager
transformers>=4.30.0
torch
Pillow
torchvision